BACKPORT: drm/i915/bdw: Expand FADD to 64bit
For error state, like the recent modification to ACTHD, FADD also gets
an upper dword. This is useful for debug to make sure the fetch address
and head are similar.


From v3.16

BUG=None
TEST=No graphics autotest regressions
TEST=Manual: no user experience regressions

Conflicts (adapt to pre-362b8af per ring error state):
	drivers/gpu/drm/i915/i915_drv.h
	drivers/gpu/drm/i915/i915_gpu_error.c

(cherry picked from commit 13ffadd1f9fa2e19ecb054b063887dcd62dde2be)
Reviewed-on: https://chromium-review.googlesource.com/216305
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Commit-Queue: James Ausmus <james.ausmus@intel.com>

+++ drivers/gpu/drm/i915/i915_drv.h
@@ -328,7 +328,7 @@ struct drm_i915_error_state {
 	u64 bbaddr[I915_NUM_RINGS];
 	u32 fault_reg[I915_NUM_RINGS];
 	u32 done_reg;
-	u32 faddr[I915_NUM_RINGS];
+	u64 faddr[I915_NUM_RINGS];
 	u64 fence[I915_MAX_NUM_FENCES];
 	struct timeval time;
 	struct drm_i915_error_ring {
